more julian code: * multiple dispatch for cover * split algorithms into multiple files * bugfixes
