+++ test/LocalPreferences.toml
@@ -1,2 +1,2 @@
 [GAIO]
-precompile_workload = false
+precompile_workload = true
